+++ server/src/constants.ts
@@ -1,3 +1,5 @@
+export const PLUGIN_VERSION = '0.2.1';
+
 export const OAUTH_CLIENT_ID = '1087611320734-u0b64f1votg3135lcc80edorfa7pg3sm.apps.googleusercontent.com';
 export const OAUTH_CLIENT_SECRET = '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';